+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -41,5 +41,5 @@ public interface FilmStorage {
 
     Collection<Film> getPopularFilms(int count);
 
-    Collection<Film> getRecommendedFilms(int userId);
+    Collection<Film> getRecommendedFilms(Long userId);
 }
